requirements.txt
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
lightning==2.5.1
pyyaml
matplotlib
tensorboard
transformers==4.52.0
